requirements.txt
# APN Core v2.0 - Minimal Dependencies

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
cryptography>=41.0.0
nats-py>=2.6.0
httpx>=0.25.0

# GUI dependencies
PyQt6>=6.6.0

# System monitoring
psutil>=5.9.0

# Database
aiosqlite>=0.19.0

# Utilities
python-dotenv>=1.0.0
